mod print
add background save, require parallel toolbox

+++ generalFunc/print_tanaka.m
@@ -1,10 +1,20 @@
-function print_tanaka(figurehandle, filename)
+function print_tanaka(figurehandle, filename, bgs)
 %print_tanaka 普段使いの一括出力
-%   1にfigure handle, 2にファイルネームの頭, 3にカットの有無
-%   png pdf(イラレ) figをコピー、pngは端を切り落とす。
-print(figurehandle, strcat(filename,'.png'), "-dpng", "-r250");
-print(figurehandle, strcat(filename,'.pdf'), "-dpdf", "-painters");
-savefig(figurehandle, strcat(filename, '.fig'), 'compact');
+%   1にfigure handle, 2にファイルネームの頭, 3にbackground or not
+%   png pdf(イラレ) figをコピー、
+if exist("bgs","var")
+    if(bgs)
+        parfeval(@print,figurehandle, strcat(filename,'.png'), "-dpng", "-r250");
+        parfeval(@print,figurehandle, strcat(filename,'.pdf'), "-dpdf", "-painters");
+        parfeval(@savefig,figurehandle, strcat(filename, '.fig'), 'compact');
+        
+    end
+else
+    print(figurehandle, strcat(filename,'.png'), "-dpng", "-r250");
+    print(figurehandle, strcat(filename,'.pdf'), "-dpdf", "-painters");
+    savefig(figurehandle, strcat(filename, '.fig'), 'compact');
+end
+
 
 end
